apps/server/requirements.txt
django
django-ninja
django-ninja-extra
rav
django-ninja-jwt[crypto]
pydantic[email]
django-cors-headers